remove stupid long long typedef

+++ src/l2_subset_independent.c
@@ -6,7 +6,7 @@ int main(int argc, char *argv[]) {
     struct input_data data;
     struct weights *w = read_point_file(&data, argc, argv);
     srand(data.seed);
-    for (ll i = 0; i < data.n_trials; i++) {
+    for (long long i = 0; i < data.n_trials; i++) {
         printf("Trial %lld\n", i);
         select_random_points(w);
         struct analytics *a = main_loop(w);
